pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clinkz"
version = "0.1.0"
description = "Autonomous AI penetration testing agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "google-generativeai>=0.5.0",
    "ollama>=0.2.0",
    "langgraph>=0.1.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
clinkz = "clinkz.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/clinkz"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["clinkz"]
